predict_app/requirements.txt
flask
Flask-API
marshmallow
python-dotenv
pymongo
pytest-flask
coverage
waitress
paste
sentry-sdk[flask]